feat(storybook): DRY ThemeProvider decorator
Allow storybook to provide a ThemeProvider with knobs attached to
customize themes on any story.
